+++ BlazorCore/Models/BaseRazorIdEntity.cs
@@ -14,8 +14,8 @@ namespace BlazorCore.Models
     {
         #region Public and private fields and properties
 
-        [Parameter] public IBaseIdEntity IdItem { get => (BaseIdEntity)Item; set => SetItem(value); }
-        [Parameter] public int Id { get => IdItem == null ? 0 : IdItem.Id; set => _ = value; }
+        [Parameter] public int Id {  get; set; }
+        public IBaseIdEntity IdItem { get => (BaseIdEntity)Item; set => SetItem(value); }
 
         #endregion
 
@@ -26,6 +26,16 @@ namespace BlazorCore.Models
         #endregion
 
         #region Public and private methods
+
+        public void SetItem(IBaseIdEntity item)
+        {
+            SetItem((IBaseEntity)item);
+        }
+
+        public void SetParentItem(IBaseIdEntity parentItem)
+        {
+            SetParentItem((IBaseEntity)parentItem);
+        }
 
         public async Task ItemSelectAsync(IBaseIdEntity item)
         {