+++ BlazorDeviceControl/Shared/Component/ItemDates.razor.cs
@@ -1,9 +1,15 @@
 ﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
+using BlazorCore;
 using BlazorCore.DAL;
 using BlazorCore.DAL.TableModels;
+using BlazorCore.Models;
+using BlazorCore.Utils;
 using Microsoft.AspNetCore.Components;
+using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace BlazorDeviceControl.Shared.Component
 {
@@ -11,41 +17,30 @@ namespace BlazorDeviceControl.Shared.Component
     {
         #region Public and private fields and properties
 
-        public string DtCreate { get; private set; }
-        public string DtModify { get; private set; }
-        private IBaseEntity _dtItem;
-        [Parameter] public IBaseEntity DtItem
-        {
-            get => _dtItem;
-            set
-            {
-                _dtItem = value;
-                DtCreate = string.Empty;
-                DtModify = string.Empty;
-
-                if (_dtItem != null && _dtItem is ScaleEntity scalesItem)
-                {
-                    DtCreate = scalesItem.CreateDate.ToString();
-                    DtModify = scalesItem.ModifiedDate.ToString();
-                }
-                else if (_dtItem != null && _dtItem is PluEntity pluEntity)
-                {
-                    DtCreate = pluEntity.CreateDate.ToString();
-                    DtModify = pluEntity.ModifiedDate.ToString();
-                }
-                else if (_dtItem != null && _dtItem is PrinterEntity printerItem)
-                {
-                    DtCreate = printerItem.CreateDate.ToString();
-                    DtModify = printerItem.ModifiedDate.ToString();
-                }
-            }
-        }
+        [Parameter] public int Id { get; set; }
+        [Parameter] public Guid Uid { get; set; }
+        [Parameter] public string DtCreate { get; set; }
+        [Parameter] public string DtModify { get; set; }
 
         #endregion
 
         #region Public and private methods
 
-        //
+        public override async Task SetParametersAsync(ParameterView parameters)
+        {
+            await base.SetParametersAsync(parameters).ConfigureAwait(true);
+            RunTasks($"{LocalizationStrings.DeviceControl.Method} {nameof(SetParametersAsync)}", "", LocalizationStrings.Share.DialogResultFail, "",
+                new List<Task> {
+                    new(async() => {
+                        if (Item is ScaleEntity scaleEntity)
+                        {
+                            DtCreate = scaleEntity.CreateDate.ToString();
+                            DtModify = scaleEntity.ModifiedDate.ToString();
+                        }
+                        await GuiRefreshWithWaitAsync();
+                    }),
+                }, true);
+        }
 
         #endregion
     }