+++ BlazorDeviceControl/Shared/Section/Nomenclatures.razor.cs
@@ -29,7 +29,7 @@ namespace BlazorDeviceControl.Shared.Section
             RunTasks($"{LocalizationStrings.DeviceControl.Method} {nameof(SetParametersAsync)}", "", LocalizationStrings.Share.DialogResultFail, "",
                 new List<Task> {
                     new(async() => {
-                        SetTable(new TableScaleEntity(BlazorCore.EnumTableScale.Printers));
+                        Table = new TableScaleEntity(EnumTableScale.Printers);
                         IdItem = null;
                         Items = null;
                         ItemsCount = 0;