+++ BlazorDeviceControl/Shared/Section/Printers.razor.cs
@@ -31,7 +31,7 @@ namespace BlazorDeviceControl.Shared.Section
             RunTasks($"{LocalizationStrings.DeviceControl.Method} {nameof(SetParametersAsync)}", "", LocalizationStrings.Share.DialogResultFail, "",
                 new List<Task> {
                     new(async() => {
-                        SetTable(new TableScaleEntity(BlazorCore.EnumTableScale.Printers));
+                        Table = new TableScaleEntity(EnumTableScale.Printers);
                         IdItem = null;
                         Items = null;
                         TemplateCategories = null;